feat: add throttle function
