+++ navigation/pure_pursuit/setup.py
@@ -1,0 +1,9 @@
+"""
+Install the lidar library
+"""
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+d = generate_distutils_setup(packages=["pure_pursuit"], package_dir={"": "src"})
+
+setup(**d)
